+++ project/src/pages/main/map.tsx
@@ -1,0 +1,7 @@
+function Map(): JSX.Element {
+  return (
+    <section className="cities__map map"></section>
+  );
+}
+
+export default Map;
